feat: 메인 페이지 애니메이션 추가
