+++ Targets/D3N/Sources/Feature/Onboarding/OnboardingNavigationStackView.swift
@@ -24,9 +24,6 @@ public struct OnboardingNavigationStackView: View {
         ) {
             WithViewStore(self.store, observe: { $0 }) { viewStore in
                 OnboardingSignUpView(store: self.store.scope(state: \.signUp, action: OnboardingNavigationStackStore.Action.signUp))
-                    .onAppear {
-                        viewStore.send(.onAppear)
-                    }
                     .navigationTitle("")
             }
         } destination: {
