+++ Targets/D3N/Sources/Feature/Today/Main/TodayMainView.swift
@@ -31,12 +31,6 @@ public struct TodayMainView: View {
                 VStack {
                     todayNewsView(viewStore: viewStore)
                         .padding()
-                    
-                    LazyVGrid(columns: .init(repeating: .init(), count: 5)) {
-                        ForEach(NewsField.allCases, id: \.self) { newsField in
-                            newsField.icon
-                        }
-                    }
                 }
             }
             .task {
@@ -47,7 +41,7 @@ public struct TodayMainView: View {
     }
     
     private func todayNewsView(viewStore: ViewStoreOf<TodayMainStore>) -> some View {
-        VStack(alignment: .leading) {
+        VStack(alignment: .leading, spacing: .zero) {
             VStack(alignment: .leading){
                 Text("최신 뉴스를 가져왔어요")
                     .font(.subheadline)
@@ -67,82 +61,22 @@ public struct TodayMainView: View {
                     )
                 }
             }
+            .padding([.horizontal, .top])
             .onTapGesture {
                 viewStore.send(.allNewsButtonTapped)
                 isPressing = false
             }
-            ForEachStore(self.store.scope(state: \.todayListItems, action: TodayMainStore.Action.todayListItems(id:action:))) {
-                TodayListItemCellView(store: $0)
+            
+            VStack(alignment: .leading, spacing: .zero) {
+                ForEachStore(self.store.scope(state: \.todayListItems, action: TodayMainStore.Action.todayListItems(id:action:))) {
+                    TodayListItemCellView(store: $0)
+                }
             }
+            .padding(.bottom, 10)
         }
-        .padding()
         .background(Color.background)
         .cornerRadius(20)
         .clipped()
         .shadow(color: .systemGray5, radius: 20)
-    }
-}
-
-struct ScrollViewGestureButtonStyle: ButtonStyle {
-    
-    init(
-        pressAction: @escaping () -> Void,
-        doubleTapTimeoutout: TimeInterval,
-        doubleTapAction: @escaping () -> Void,
-        longPressTime: TimeInterval,
-        longPressAction: @escaping () -> Void,
-        endAction: @escaping () -> Void
-    ) {
-        self.pressAction = pressAction
-        self.doubleTapTimeoutout = doubleTapTimeoutout
-        self.doubleTapAction = doubleTapAction
-        self.longPressTime = longPressTime
-        self.longPressAction = longPressAction
-        self.endAction = endAction
-    }
-    
-    private var doubleTapTimeoutout: TimeInterval
-    private var longPressTime: TimeInterval
-    
-    private var pressAction: () -> Void
-    private var longPressAction: () -> Void
-    private var doubleTapAction: () -> Void
-    private var endAction: () -> Void
-    
-    @State
-    var doubleTapDate = Date()
-    
-    @State
-    var longPressDate = Date()
-    
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .onChange(of: configuration.isPressed) { isPressed in
-                longPressDate = Date()
-                if isPressed {
-                    pressAction()
-                    doubleTapDate = tryTriggerDoubleTap() ? .distantPast : .now
-                    tryTriggerLongPressAfterDelay(triggered: longPressDate)
-                } else {
-                    endAction()
-                }
-            }
-    }
-}
-
-private extension ScrollViewGestureButtonStyle {
-    
-    func tryTriggerDoubleTap() -> Bool {
-        let interval = Date().timeIntervalSince(doubleTapDate)
-        guard interval < doubleTapTimeoutout else { return false }
-        doubleTapAction()
-        return true
-    }
-    
-    func tryTriggerLongPressAfterDelay(triggered date: Date) {
-        DispatchQueue.main.asyncAfter(deadline: .now() + longPressTime) {
-            guard date == longPressDate else { return }
-            longPressAction()
-        }
     }
 }
